# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Clausly API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type DocumentDocumentType =
  (typeof DocumentDocumentType)[keyof typeof DocumentDocumentType];

export const DocumentDocumentType = {
  nda: "nda",
  privacy_policy: "privacy_policy",
  contractor_agreement: "contractor_agreement",
  terms_of_service: "terms_of_service",
} as const;

export interface Document {
  id: number;
  userId: string;
  documentType: DocumentDocumentType;
  title: string;
  partyA: string;
  /** @nullable */
  partyB: string | null;
  content: string;
  /** @nullable */
  metadata: string | null;
  createdAt: string;
  updatedAt: string;
}

export type CreateDocumentBodyDocumentType =
  (typeof CreateDocumentBodyDocumentType)[keyof typeof CreateDocumentBodyDocumentType];

export const CreateDocumentBodyDocumentType = {
  nda: "nda",
  privacy_policy: "privacy_policy",
  contractor_agreement: "contractor_agreement",
  terms_of_service: "terms_of_service",
} as const;

export interface CreateDocumentBody {
  documentType: CreateDocumentBodyDocumentType;
  partyA: string;
  partyB?: string;
  keyTerms: string;
  jurisdiction?: string;
  additionalContext?: string;
}

export interface ExplainClauseBody {
  clauseText: string;
}

export interface ExplainClauseResponse {
  explanation: string;
  whyItMatters: string;
  favoredParty: string;
}

export interface Review {
  id: number;
  userId: string;
  title: string;
  contractText: string;
  riskScore: number;
  /** @nullable */
  riskyClausesJson: string | null;
  /** @nullable */
  summaryJson: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateReviewBody {
  contractText: string;
  title: string;
}

export interface Comparison {
  id: number;
  userId: string;
  title: string;
  contractAText: string;
  contractBText: string;
  /** @nullable */
  differencesJson: string | null;
  /** @nullable */
  missingClausesJson: string | null;
  /** @nullable */
  assessmentJson: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateComparisonBody {
  title: string;
  contractAText: string;
  contractBText: string;
}

export type TemplateDocumentType =
  (typeof TemplateDocumentType)[keyof typeof TemplateDocumentType];

export const TemplateDocumentType = {
  nda: "nda",
  privacy_policy: "privacy_policy",
  contractor_agreement: "contractor_agreement",
  terms_of_service: "terms_of_service",
} as const;

export interface Template {
  id: number;
  userId: string;
  name: string;
  documentType: TemplateDocumentType;
  partyA: string;
  /** @nullable */
  jurisdiction: string | null;
  keyTerms: string;
  /** @nullable */
  additionalContext: string | null;
  createdAt: string;
}

export type CreateTemplateBodyDocumentType =
  (typeof CreateTemplateBodyDocumentType)[keyof typeof CreateTemplateBodyDocumentType];

export const CreateTemplateBodyDocumentType = {
  nda: "nda",
  privacy_policy: "privacy_policy",
  contractor_agreement: "contractor_agreement",
  terms_of_service: "terms_of_service",
} as const;

export interface CreateTemplateBody {
  name: string;
  documentType: CreateTemplateBodyDocumentType;
  partyA: string;
  jurisdiction?: string;
  keyTerms: string;
  additionalContext?: string;
}

export type DashboardStatsDocumentsByType = { [key: string]: number };

export interface DashboardStats {
  totalDocuments: number;
  totalReviews: number;
  templateCount: number;
  documentsByType: DashboardStatsDocumentsByType;
  /** @nullable */
  avgRiskScore: number | null;
}

export interface RecentActivity {
  documents: Document[];
  reviews: Review[];
}