# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Clausly API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: documents
    description: Document operations
  - name: reviews
    description: Contract review operations
  - name: dashboard
    description: Dashboard operations
  - name: comparisons
    description: Document comparison operations
  - name: templates
    description: Document template operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /documents:
    get:
      operationId: listDocuments
      tags: [documents]
      summary: List user's generated documents
      responses:
        "200":
          description: List of documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Document"
        "401":
          description: Unauthorized
    post:
      operationId: createDocument
      tags: [documents]
      summary: Generate a legal document using AI
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDocumentBody"
      responses:
        "201":
          description: Created document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
        "400":
          description: Bad request
        "401":
          description: Unauthorized

  /documents/{id}:
    get:
      operationId: getDocument
      tags: [documents]
      summary: Get a single document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
        "401":
          description: Unauthorized
        "404":
          description: Not found
    delete:
      operationId: deleteDocument
      tags: [documents]
      summary: Delete a document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized
        "404":
          description: Not found

  /documents/{id}/explain-clause:
    post:
      operationId: explainClause
      tags: [documents]
      summary: Get AI plain-English explanation of a document clause
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExplainClauseBody"
      responses:
        "200":
          description: Clause explanation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExplainClauseResponse"
        "400":
          description: Bad request
        "401":
          description: Unauthorized
        "404":
          description: Not found

  /reviews:
    get:
      operationId: listReviews
      tags: [reviews]
      summary: List user's contract reviews
      responses:
        "200":
          description: List of reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"
        "401":
          description: Unauthorized
    post:
      operationId: createReview
      tags: [reviews]
      summary: Analyze a contract using AI
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReviewBody"
      responses:
        "201":
          description: Created review
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
        "400":
          description: Bad request
        "401":
          description: Unauthorized

  /reviews/{id}:
    get:
      operationId: getReview
      tags: [reviews]
      summary: Get a single review
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Review
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
        "401":
          description: Unauthorized
        "404":
          description: Not found
    delete:
      operationId: deleteReview
      tags: [reviews]
      summary: Delete a review
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized
        "404":
          description: Not found

  /comparisons:
    get:
      operationId: listComparisons
      tags: [comparisons]
      summary: List user's document comparisons
      responses:
        "200":
          description: List of comparisons
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Comparison"
        "401":
          description: Unauthorized
    post:
      operationId: createComparison
      tags: [comparisons]
      summary: Compare two contracts using AI
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateComparisonBody"
      responses:
        "201":
          description: Created comparison
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comparison"
        "400":
          description: Bad request
        "401":
          description: Unauthorized

  /comparisons/{id}:
    get:
      operationId: getComparison
      tags: [comparisons]
      summary: Get a single comparison
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Comparison
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comparison"
        "401":
          description: Unauthorized
        "404":
          description: Not found
    delete:
      operationId: deleteComparison
      tags: [comparisons]
      summary: Delete a comparison
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized
        "404":
          description: Not found

  /templates:
    get:
      operationId: listTemplates
      tags: [templates]
      summary: List user's document templates
      responses:
        "200":
          description: List of templates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Template"
        "401":
          description: Unauthorized
    post:
      operationId: createTemplate
      tags: [templates]
      summary: Save a document template
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTemplateBody"
      responses:
        "201":
          description: Created template
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Template"
        "400":
          description: Bad request
        "401":
          description: Unauthorized

  /templates/{id}:
    delete:
      operationId: deleteTemplate
      tags: [templates]
      summary: Delete a template
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized
        "404":
          description: Not found

  /dashboard/stats:
    get:
      operationId: getDashboardStats
      tags: [dashboard]
      summary: Get dashboard summary stats
      responses:
        "200":
          description: Dashboard stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"
        "401":
          description: Unauthorized

  /dashboard/recent:
    get:
      operationId: getRecentActivity
      tags: [dashboard]
      summary: Get recent documents and reviews activity
      responses:
        "200":
          description: Recent activity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecentActivity"
        "401":
          description: Unauthorized

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Document:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        documentType:
          type: string
          enum: [nda, privacy_policy, contractor_agreement, terms_of_service]
        title:
          type: string
        partyA:
          type: string
        partyB:
          type: ["string", "null"]
        content:
          type: string
        metadata:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - documentType
        - title
        - partyA
        - partyB
        - content
        - metadata
        - createdAt
        - updatedAt

    CreateDocumentBody:
      type: object
      properties:
        documentType:
          type: string
          enum: [nda, privacy_policy, contractor_agreement, terms_of_service]
        partyA:
          type: string
        partyB:
          type: string
        keyTerms:
          type: string
        jurisdiction:
          type: string
        additionalContext:
          type: string
      required:
        - documentType
        - partyA
        - keyTerms

    ExplainClauseBody:
      type: object
      properties:
        clauseText:
          type: string
      required:
        - clauseText

    ExplainClauseResponse:
      type: object
      properties:
        explanation:
          type: string
        whyItMatters:
          type: string
        favoredParty:
          type: string
      required:
        - explanation
        - whyItMatters
        - favoredParty

    Review:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        title:
          type: string
        contractText:
          type: string
        riskScore:
          type: integer
        riskyClausesJson:
          type: ["string", "null"]
        summaryJson:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - title
        - contractText
        - riskScore
        - riskyClausesJson
        - summaryJson
        - createdAt
        - updatedAt

    CreateReviewBody:
      type: object
      properties:
        contractText:
          type: string
        title:
          type: string
      required:
        - contractText
        - title

    Comparison:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        title:
          type: string
        contractAText:
          type: string
        contractBText:
          type: string
        differencesJson:
          type: ["string", "null"]
        missingClausesJson:
          type: ["string", "null"]
        assessmentJson:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - title
        - contractAText
        - contractBText
        - differencesJson
        - missingClausesJson
        - assessmentJson
        - createdAt
        - updatedAt

    CreateComparisonBody:
      type: object
      properties:
        title:
          type: string
        contractAText:
          type: string
        contractBText:
          type: string
      required:
        - title
        - contractAText
        - contractBText

    Template:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        name:
          type: string
        documentType:
          type: string
          enum: [nda, privacy_policy, contractor_agreement, terms_of_service]
        partyA:
          type: string
        jurisdiction:
          type: ["string", "null"]
        keyTerms:
          type: string
        additionalContext:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - name
        - documentType
        - partyA
        - jurisdiction
        - keyTerms
        - additionalContext
        - createdAt

    CreateTemplateBody:
      type: object
      properties:
        name:
          type: string
        documentType:
          type: string
          enum: [nda, privacy_policy, contractor_agreement, terms_of_service]
        partyA:
          type: string
        jurisdiction:
          type: string
        keyTerms:
          type: string
        additionalContext:
          type: string
      required:
        - name
        - documentType
        - partyA
        - keyTerms

    DashboardStats:
      type: object
      properties:
        totalDocuments:
          type: integer
        totalReviews:
          type: integer
        templateCount:
          type: integer
        documentsByType:
          type: object
          additionalProperties:
            type: integer
        avgRiskScore:
          type: ["number", "null"]
      required:
        - totalDocuments
        - totalReviews
        - templateCount
        - documentsByType
        - avgRiskScore

    RecentActivity:
      type: object
      properties:
        documents:
          type: array
          items:
            $ref: "#/components/schemas/Document"
        reviews:
          type: array
          items:
            $ref: "#/components/schemas/Review"
      required:
        - documents
        - reviews