# lib/api-zod/src/generated/types/comparison.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Clausly API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Comparison {
  id: number;
  userId: string;
  title: string;
  contractAText: string;
  contractBText: string;
  /** @nullable */
  differencesJson: string | null;
  /** @nullable */
  missingClausesJson: string | null;
  /** @nullable */
  assessmentJson: string | null;
  createdAt: Date;
  updatedAt: Date;
}