# lib/api-zod/src/generated/types/createComparisonBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Clausly API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateComparisonBody {
  title: string;
  contractAText: string;
  contractBText: string;
}