# lib/api-zod/src/generated/types/createTemplateBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Clausly API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreateTemplateBodyDocumentType } from "./createTemplateBodyDocumentType";

export interface CreateTemplateBody {
  name: string;
  documentType: CreateTemplateBodyDocumentType;
  partyA: string;
  jurisdiction?: string;
  keyTerms: string;
  additionalContext?: string;
}