# lib/api-zod/src/generated/types/dashboardStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Clausly API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DashboardStatsDocumentsByType } from "./dashboardStatsDocumentsByType";

export interface DashboardStats {
  totalDocuments: number;
  totalReviews: number;
  templateCount: number;
  documentsByType: DashboardStatsDocumentsByType;
  /** @nullable */
  avgRiskScore: number | null;
}