# lib/api-zod/src/generated/types/template.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Clausly API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TemplateDocumentType } from "./templateDocumentType";

export interface Template {
  id: number;
  userId: string;
  name: string;
  documentType: TemplateDocumentType;
  partyA: string;
  /** @nullable */
  jurisdiction: string | null;
  keyTerms: string;
  /** @nullable */
  additionalContext: string | null;
  createdAt: Date;
}